requirements.txt
py-cord
python-dotenv
dataclasses-json